create .debug.html files which can be ran directly in a browser for debugging
