+++ build_test/build.yaml
@@ -2,14 +2,15 @@ builders:
   test_bootstrap:
     target: "build_test"
     import: "package:build_test/builder.dart"
-    builder_factories: ["testBootstrapBuilder"]
+    builder_factories: ["debugTestBuilder", "testBootstrapBuilder"]
     build_extensions:
       _test.dart:
         - _test.dart.vm_test.dart
         - _test.dart.browser_test.dart
         - _test.dart.node_test.dart
         - _test.html
-    is_optional: True
+        - _test.debug.html
+    is_optional: False
     build_to: cache
     auto_apply: root_package
     defaults:
